Replace OAuth and OpenID function from utils.py to __init__.py near blueprints
