+++ wedapp/auth/models.py
@@ -14,7 +14,6 @@ class User(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     username = db.Column(db.String(255), nullable=False, unique=True)
     password = db.Column(db.String(255))
-    # email = db.Column(db.String(255), nullable=False, unique=True)
     posts = db.relationship("Post", backref="users", lazy="dynamic")
 
     # def __init__(self, username):
